+++ tala/model/grammar/intent.py
@@ -71,8 +71,3 @@ class Answer(Intent):
 class AnswerNegation(Intent):
     def __init__(self, text_chunks, required_entities):
         super(AnswerNegation, self).__init__("answer_negation", text_chunks, required_entities)
-
-
-class Negative(Intent):
-    def __init__(self):
-        super(Negative, self).__init__("negative", [""], [])